:sparkles: refactor : change dto dir apis to domains add controller and service
