+++ src/test/java/com/depromeet/streetdrop/api/area/village/controller/VillageItemControllerTest.java
@@ -2,7 +2,7 @@ package com.depromeet.streetdrop.api.area.village.controller;
 
 import com.depromeet.streetdrop.domains.area.village.service.VillageItemService;
 import com.depromeet.streetdrop.apis.area.village.controller.VillageItemController;
-import com.depromeet.streetdrop.apis.area.village.dto.response.VillageItemsCountResponseDto;
+import com.depromeet.streetdrop.domains.area.village.dto.response.VillageItemsCountResponseDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
